add sign for user init DB
