+++ cmd/meal-planner/main.go
@@ -2,9 +2,8 @@ package main
 
 import (
 	"context"
-	"encoding/json"
-	"fmt"
 	"github.com/go-chi/chi/v5"
+	"health-food-for-each/internal"
 	"log"
 	"net/http"
 	"os"
@@ -13,66 +12,11 @@ import (
 	"time"
 )
 
-// Авторизация - принимаем на вход json: login - password. Парсим его, хэшируем пароль и записываем в БД. Ответ - ОК
-func AuthPage(w http.ResponseWriter, r *http.Request) {
-	decoder := json.NewDecoder(r.Body)
-	var req Request
-	err := decoder.Decode(&req)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusBadRequest)
-		return
-	}
-
-	link := req.URL
-	shortURL := fmt.Sprintf("%s/%s", Config.BaseURL, id)
-
-	if app.Cfg.DatabaseAddr != "" {
-		err := app.Cfg.URLMapDB.SetDB(context.Background(), id, link, app.Cfg.UserID)
-		if err != nil {
-			originalURL, err := app.Cfg.URLMapDB.GetReverse(context.Background(), link, app.Cfg.UserID)
-			if err != nil {
-				log.Println(err)
-				return
-			}
-			shortURL := fmt.Sprintf("%s/%s", app.Cfg.BaseURL, originalURL)
-			resp := Response{Result: shortURL}
-			responseData, err := json.Marshal(resp)
-			if err != nil {
-				return
-			}
-
-			w.Header().Set("Content-Type", "application/json")
-			w.WriteHeader(http.StatusConflict)
-			w.Write(responseData)
-			return
-		}
-	} else {
-		app.Cfg.URLMap.Set(id, link)
-	}
-
-	resp := Response{Result: shortURL}
-	responseData, err := json.Marshal(resp)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusCreated)
-	w.Write(responseData)
-}
-
-type Request struct {
-	URL string `json:"url"`
-}
-
-type Response struct {
-	Result string `json:"result"`
-}
+var ServerAddress string
 
 func main() {
-	SetFlags()
-	SetConfig()
+	//SetFlags()
+	//SetConfig()
 
 	//Подключение к БД
 	//var once sync.Once
@@ -87,19 +31,22 @@ func main() {
 	//r.Use(app.AuthMiddleware)
 
 	//Хэндлеры
-	r.Get("/", MainPage)
-	r.Get("/auth", AuthPage)
-	r.Get("/info", InfoPage)
-	r.Get("/{id}", Redirect)
-	r.Get("/ping", BDConnection)
-	r.Get("/settings", Settings)
-	r.Get("/refresh", Refresh)
-	r.Get("/my-food", ShowMyFood)
-	r.Post("/my-food", AddToMyFood)
-	r.Delete("/my-food", RemoveMyFood)
+	//r.Get("/", MainPage)
+	r.Get("/auth", internal.SignIn)
+	r.Get("/auth", internal.SignUp)
+	//r.Get("/info", InfoPage)
+	//r.Get("/{id}", Redirect)
+	//r.Get("/ping", BDConnection)
+	//r.Get("/settings", Settings)
+	//r.Get("/refresh", Refresh)
+	//r.Get("/my-food", ShowMyFood)
+	//r.Post("/my-food", AddToMyFood)
+	//r.Delete("/my-food", RemoveMyFood)
 
+	//TODO: вынести в конфиг
+	ServerAddress = "http://localhost:8080"
 	srv := &http.Server{
-		Addr:    Config.ServerAddr,
+		Addr:    ServerAddress,
 		Handler: r,
 	}
 
@@ -119,7 +66,7 @@ func main() {
 		}
 	}()
 
-	log.Printf("Server is listening on %s\n", Config.ServerAddr)
+	log.Printf("Server is listening on %s\n", ServerAddress)
 
 	if err := srv.ListenAndServe(); err != http.ErrServerClosed {
 		log.Fatalf("HTTP server ListenAndServe: %v", err)
