+++ internal/dataBase.go
@@ -1,0 +1,7 @@
+package internal
+
+type DataBase interface {
+	SetURL
+	GetURL
+	SaveToFile() error
+}
